requirements.txt
Flask
ftfy
Pillow
pytesseract
